# perf(serialize): faster symbol and number comparison