--- a/src/serialize.ts
+++ b/src/serialize.ts
@@ -11,32 +11,23 @@
  * @return {string} serialized string value
  */
 export function serialize(input: any): string {
-  return _serialize(input);
-}
-
-function _serialize(input: any, context?: Map<any, string>): string {
   if (typeof input === "string") {
     return `'${input}'`;
   }
-  const serializer = new Serializer(context);
-  return serializer.serialize(input);
+  return new Serializer().serialize(input);
 }
 
 const Serializer = /*@__PURE__*/ (function () {
   class Serializer {
-    #context: Map<any, string>;
-
-    constructor(context = new Map()) {
-      this.#context = context;
-    }
+    #context = new Map();
 
     compare(a: any, b: any): number {
       // Uses fast path to compare primitive values (string, number, bigint, boolean, null, undefined)
       // Only symbol, function and object values need to be full serialized
       // TODO: we could make it faster by fast path if both sides are numbers (need benchmarks first)
       return String.prototype.localeCompare.call(
-        toComparableString(a) ?? _serialize(a, this.#context),
-        toComparableString(b) ?? _serialize(b, this.#context),
+        toComparableString(a) ?? this.serialize(a),
+        toComparableString(b) ?? this.serialize(b),
       );
     }